// src/app/editor/components/editor/TextSticker/types.ts
export type TTextStickerData = {
  title: string;
  titleColor: "#3F28C6";
  subtitle: string;
  description: string;
  images: Array<Record<string, string>>;
  editorImages: Array<Record<string, string>>;
  stickers: string[];
  sortIndex: number;
  stickerBgColor: string;
  stickerTextColor: string;
};